Makefile: Use GnuPG to verify the Bintray JARs
This commit changes from the hard-coded hash over to OpenPGP [1]
verification for the tool JAR.  The key in goneall.gpg is
0xEA760AF7CBD0E4F9013F06FB8CCBE63DE91BFF1C, imported with:

  $ gpg --search 0x8CCBE63DE91BFF1C
  gpg: data source: http://5.45.108.219:11371
  (1)     Gary O'Neall (Key for signing jar files) <gary@sourceauditor.com>
            4096 bit RSA key 8CCBE63DE91BFF1C, created: 2017-12-13
  Keys 1-1 of 1 for "0x8CCBE63DE91BFF1C".  Enter number(s), N)ext, or Q)uit > 1
  gpg: key 8CCBE63DE91BFF1C: public key "Gary O'Neall (Key for signing jar files) <gary@sourceauditor.com>" imported
  gpg: Total number processed: 1
  gpg:               imported: 1

The source keyserver is from hkp://pool.sks-keyservers.net.  I then
exported the key with:

  $ gpg --export 0x8CCBE63DE91BFF1C >goneall.gpg

When exported without ASCII armor, such keys can be used as a keyring.
The --no-default-keyring option disables the user's default keyring
(if any.  There probably won't be one on Travis).  And the --keyring
./goneall.gpg slots in our local file.  We need the ./ (at least with
GnuPG 2.1.20), because gpg2(1) has:

  If the filename does not contain a slash, it is assumed to be in the
  GnuPG home directory ("~/.gnupg" if --homedir or $GNUPGHOME is not
  used).

The output may still include:

  gpg: WARNING: This key is not certified with a trusted signature!
  gpg:          There is no indication that the signature belongs to the owner.

but that's ok, because gpg still exits zero.  And because of our
keyring manipulation, we know the key is trusted.  Signatures from any
other key will instead show:

  gpg: Can't check signature: No public key

and exit 2.

[1]: https://tools.ietf.org/html/rfc4880